FIX : OSB crash on remove + upload
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@14348 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ src/ImportExport/ImportExportOsmBin.cpp
@@ -129,21 +129,17 @@ bool OsbRegion::clearTile(qint32 tile, MapDocument* d, OsbMapLayer* theLayer)
 
 	for (qint32 i=0; i<theOsb->theTileIndex[tile].size(); ++i) {
 		MapFeature* F = theOsb->theTileIndex[tile][i];
-		if (F && theLayer->featRefCount.contains(F)) {
-			if (F->layer() == theLayer) {
-				int theCount = --theLayer->featRefCount[F];
-				if (!(theCount)) {
-					quint32 j;
-					for (j=0; j<F->sizeParents(); ++j)
-						if (F->getParent(j)->layer() != theLayer)
-							break;
-					if (j == F->sizeParents()) {
-						theLayer->deleteFeature(F);
-						theLayer->featRefCount.remove(F);
-					}
+		if (F && theLayer->exists(F) && theLayer->featRefCount.contains(F)) {
+			int theCount = --theLayer->featRefCount[F];
+			if (!(theCount)) {
+				quint32 j;
+				for (j=0; j<F->sizeParents(); ++j)
+					if (F->getParent(j)->layer() != theLayer)
+						break;
+				if (j == F->sizeParents()) {
+					theLayer->deleteFeature(F);
+					theLayer->featRefCount.remove(F);
 				}
-			} else {
-				theLayer->featRefCount.remove(F);
 			}
 		} else {
 			theLayer->featRefCount.remove(F);
